+++ lib/Configuration.php
@@ -100,7 +100,7 @@ class Configuration {
 	 *
 	 * @var string
 	 */
-	protected $userAgent = 'BHR-SDK/2.0.2/PHP';
+	protected $userAgent = 'BHR-SDK/2.1.0/PHP';
 
 	/**
 	 * Debug switch (default set to false)
@@ -481,7 +481,7 @@ class Configuration {
 		$report .= '    OS: ' . php_uname() . PHP_EOL;
 		$report .= '    PHP Version: ' . PHP_VERSION . PHP_EOL;
 		$report .= '    The version of the OpenAPI document: 1.0' . PHP_EOL;
-		$report .= '    SDK Package Version: 2.0.2' . PHP_EOL;
+		$report .= '    SDK Package Version: 2.1.0' . PHP_EOL;
 		$report .= '    Temp Folder Path: ' . self::getDefaultConfiguration()->getTempFolderPath() . PHP_EOL;
 
 		return $report;